+++ domain/cart.go
@@ -19,18 +19,18 @@ type CartController interface {
 
 type CartControllerPayloadCreateCartItem struct {
 	ProductUID string `json:"product_uid"`
-	Quantity   uint64 `json:"quantity"`
+	Quantity   int    `json:"quantity"`
 }
 
 type CartControllerPayloadUpdateCartItem struct {
-	Quantity uint64 `json:"quantity"`
+	Quantity int `json:"quantity"`
 }
 
 type CartControllerResponseGetCart struct {
 	UID              string                               `db:"uid" json:"uid"`
-	Quantity         uint64                               `db:"quantity" json:"quantity"`
+	Quantity         int                                  `db:"quantity" json:"quantity"`
 	TotalPrice       string                               `db:"total_price" json:"total_price"`
-	TotalPriceValue  uint64                               `db:"total_price_value" json:"total_price_value"`
+	TotalPriceValue  int                                  `db:"total_price_value" json:"total_price_value"`
 	TotalWeight      string                               `db:"total_weight" json:"total_weight"`
 	TotalWeightValue float64                              `db:"total_weight_value" json:"total_weight_value"`
 	CartItems        []ControllerResponsePropertyCartItem `db:"cart_items" json:"cart_items"`
@@ -38,9 +38,9 @@ type CartControllerResponseGetCart struct {
 
 type ControllerResponsePropertyCartItem struct {
 	UID              string  `db:"uid" json:"uid"`
-	Quantity         uint64  `db:"quantity" json:"quantity"`
+	Quantity         int     `db:"quantity" json:"quantity"`
 	TotalPrice       string  `db:"total_price" json:"total_price"`
-	TotalPriceValue  uint64  `db:"total_price_value" json:"total_price_value"`
+	TotalPriceValue  int     `db:"total_price_value" json:"total_price_value"`
 	TotalWeight      string  `db:"total_weight" json:"total_weight"`
 	TotalWeightValue float64 `db:"total_weight_value" json:"total_weight_value"`
 
@@ -51,22 +51,22 @@ type ControllerResponsePropertyCartItem struct {
 	ProductWeight      string  `db:"product_weight" json:"product_weight"`
 	ProductWeightValue float64 `db:"product_weight_value" json:"product_weight_value"`
 	BasePrice          string  `db:"base_price" json:"base_price"`
-	BasePriceValue     uint64  `db:"base_price_value" json:"base_price_value"`
+	BasePriceValue     int     `db:"base_price_value" json:"base_price_value"`
 	OfferPrice         string  `db:"offer_price" json:"offer_price"`
-	OfferPriceValue    uint64  `db:"offer_price_value" json:"offer_price_value"`
+	OfferPriceValue    int     `db:"offer_price_value" json:"offer_price_value"`
 	Discount           uint8   `db:"discount" json:"discount"`
 }
 
 // Usecase
 type CartUsecase interface {
 	// Cart
-	GetCartByUserID(userID uint64) (*CartControllerResponseGetCart, error)
-	GetCartByUserIDMiddleware(userID uint64) (*CartModel, error)
+	GetCartByUserID(userID int) (*CartControllerResponseGetCart, error)
+	GetCartByUserIDMiddleware(userID int) (*CartModel, error)
 
 	// Cart item
 	CreateCartItem(payload *CartUsecasePayloadCreateCartItem) (string, error)
 	GetCartItemByUID(UID string) (*CartItemModel, error)
-	GetCartItemByProductID(productID uint64) (*CartItemModel, error)
+	GetCartItemByProductID(productID int) (*CartItemModel, error)
 	UpdateCartItem(payload *CartUsecasePayloadUpdateCartItem) error
 	DeleteCartItemByUID(payload *CartUsecasePayloadDeleteCartItem) error
 }
@@ -74,14 +74,14 @@ type CartUsecase interface {
 type CartUsecasePayloadCreateCartItem struct {
 	Cart     *CartModel    `json:"cart"`
 	Product  *ProductModel `json:"product"`
-	Quantity uint64        `json:"quantity"`
+	Quantity int           `json:"quantity"`
 }
 
 type CartUsecasePayloadUpdateCartItem struct {
 	Cart     *CartModel     `json:"cart"`
 	CartItem *CartItemModel `json:"cart_item"`
 	UID      string         `json:"uid"`
-	Quantity uint64         `json:"quantity"`
+	Quantity int            `json:"quantity"`
 }
 
 type CartUsecasePayloadDeleteCartItem struct {
@@ -92,28 +92,28 @@ type CartUsecasePayloadDeleteCartItem struct {
 
 // Repository
 type CartModel struct {
-	ID               uint64  `db:"id" json:"id"`
+	ID               int     `db:"id" json:"id"`
 	UID              string  `db:"uid" json:"uid"`
-	Quantity         uint64  `db:"quantity" json:"quantity"`
+	Quantity         int     `db:"quantity" json:"quantity"`
 	TotalPrice       string  `db:"total_price" json:"total_price"`
-	TotalPriceValue  uint64  `db:"total_price_value" json:"total_price_value"`
+	TotalPriceValue  int     `db:"total_price_value" json:"total_price_value"`
 	TotalWeight      string  `db:"total_weight" json:"total_weight"`
 	TotalWeightValue float64 `db:"total_weight_value" json:"total_weight_value"`
 
 	// Relationship
 	CartItems []CartItemModel `db:"cart_items" json:"cart_items"`
-	UserID    uint64          `db:"user_id" json:"user_id"`
+	UserID    int             `db:"user_id" json:"user_id"`
 
 	CreatedAt time.Time `db:"created_at" json:"created_at"`
 	UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
 }
 
 type CartItemModel struct {
-	ID               uint64  `db:"id" json:"id"`
+	ID               int     `db:"id" json:"id"`
 	UID              string  `db:"uid" json:"uid"`
-	Quantity         uint64  `db:"quantity" json:"quantity"`
+	Quantity         int     `db:"quantity" json:"quantity"`
 	TotalPrice       string  `db:"total_price" json:"total_price"`
-	TotalPriceValue  uint64  `db:"total_price_value" json:"total_price_value"`
+	TotalPriceValue  int     `db:"total_price_value" json:"total_price_value"`
 	TotalWeight      string  `db:"total_weight" json:"total_weight"`
 	TotalWeightValue float64 `db:"total_weight_value" json:"total_weight_value"`
 
@@ -124,14 +124,14 @@ type CartItemModel struct {
 	ProductWeight      string  `db:"product_weight" json:"product_weight"`
 	ProductWeightValue float64 `db:"product_weight_value" json:"product_weight_value"`
 	BasePrice          string  `db:"base_price" json:"base_price"`
-	BasePriceValue     uint64  `db:"base_price_value" json:"base_price_value"`
+	BasePriceValue     int     `db:"base_price_value" json:"base_price_value"`
 	OfferPrice         string  `db:"offer_price" json:"offer_price"`
-	OfferPriceValue    uint64  `db:"offer_price_value" json:"offer_price_value"`
+	OfferPriceValue    int     `db:"offer_price_value" json:"offer_price_value"`
 	Discount           uint8   `db:"discount" json:"discount"`
 
 	// Relationship
-	CartID    uint64 `db:"cart_id" json:"cart_id"`
-	ProductID uint64 `db:"product_id" json:"product_id"`
+	CartID    int `db:"cart_id" json:"cart_id"`
+	ProductID int `db:"product_id" json:"product_id"`
 
 	CreatedAt time.Time `db:"created_at" json:"created_at"`
 	UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
@@ -143,21 +143,21 @@ type CartRepository interface {
 	// Cart
 	CreateCart() error
 	GetCartByUID(UID string) (*CartModel, error)
-	GetCartByUserID(userID uint64) (*CartModel, error)
+	GetCartByUserID(userID int) (*CartModel, error)
 
 	// Cart item
 	CreateCartItem(cartItemPayload CartRepositoryPayloadCreateCartItem, cartPayload CartRepositoryPayloadUpdateCart) (string, error)
 	GetCartItemByUID(UID string) (*CartItemModel, error)
-	GetCartItemByProductID(productID uint64) (*CartItemModel, error)
+	GetCartItemByProductID(productID int) (*CartItemModel, error)
 	UpdateCartItem(cartItemPayload CartRepositoryPayloadUpdateCartItem, cartPayload CartRepositoryPayloadUpdateCart) error
 	DeleteCartItemByUID(UID string, cartPayload CartRepositoryPayloadUpdateCart) error
 }
 
 type CartRepositoryPayloadUpdateCart struct {
 	UID              string  `db:"uid" json:"uid"`
-	Quantity         uint64  `db:"quantity" json:"quantity"`
+	Quantity         int     `db:"quantity" json:"quantity"`
 	TotalPrice       string  `db:"total_price" json:"total_price"`
-	TotalPriceValue  uint64  `db:"total_price_value" json:"total_price_value"`
+	TotalPriceValue  int     `db:"total_price_value" json:"total_price_value"`
 	TotalWeight      string  `db:"total_weight" json:"total_weight"`
 	TotalWeightValue float64 `db:"total_weight_value" json:"total_weight_value"`
 
@@ -166,9 +166,9 @@ type CartRepositoryPayloadUpdateCart struct {
 
 type CartRepositoryPayloadCreateCartItem struct {
 	UID              string  `db:"uid" json:"uid"`
-	Quantity         uint64  `db:"quantity" json:"quantity"`
+	Quantity         int     `db:"quantity" json:"quantity"`
 	TotalPrice       string  `db:"total_price" json:"total_price"`
-	TotalPriceValue  uint64  `db:"total_price_value" json:"total_price_value"`
+	TotalPriceValue  int     `db:"total_price_value" json:"total_price_value"`
 	TotalWeight      string  `db:"total_weight" json:"total_weight"`
 	TotalWeightValue float64 `db:"total_weight_value" json:"total_weight_value"`
 
@@ -179,14 +179,14 @@ type CartRepositoryPayloadCreateCartItem struct {
 	ProductWeight      string  `db:"product_weight" json:"product_weight"`
 	ProductWeightValue float64 `db:"product_weight_value" json:"product_weight_value"`
 	BasePrice          string  `db:"base_price" json:"base_price"`
-	BasePriceValue     uint64  `db:"base_price_value" json:"base_price_value"`
+	BasePriceValue     int     `db:"base_price_value" json:"base_price_value"`
 	OfferPrice         string  `db:"offer_price" json:"offer_price"`
-	OfferPriceValue    uint64  `db:"offer_price_value" json:"offer_price_value"`
-	Discount           uint8   `db:"discount" json:"discount"`
+	OfferPriceValue    int     `db:"offer_price_value" json:"offer_price_value"`
+	Discount           int     `db:"discount" json:"discount"`
 
 	// Relationship
-	CartID    uint64 `db:"cart_id" json:"cart_id"`
-	ProductID uint64 `db:"product_id" json:"product_id"`
+	CartID    int `db:"cart_id" json:"cart_id"`
+	ProductID int `db:"product_id" json:"product_id"`
 
 	CreatedAt time.Time `db:"created_at" json:"created_at"`
 	UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
@@ -194,9 +194,9 @@ type CartRepositoryPayloadCreateCartItem struct {
 
 type CartRepositoryPayloadUpdateCartItem struct {
 	UID              string  `db:"uid" json:"uid"`
-	Quantity         uint64  `db:"quantity" json:"quantity"`
+	Quantity         int     `db:"quantity" json:"quantity"`
 	TotalPrice       string  `db:"total_price" json:"total_price"`
-	TotalPriceValue  uint64  `db:"total_price_value" json:"total_price_value"`
+	TotalPriceValue  int     `db:"total_price_value" json:"total_price_value"`
 	TotalWeight      string  `db:"total_weight" json:"total_weight"`
 	TotalWeightValue float64 `db:"total_weight_value" json:"total_weight_value"`
